refactor(models) : relocate ticket models to backend
