+++ src/backend/models/__init__.py
@@ -1,0 +1,20 @@
+"""Domain models for the backend layer."""
+
+from .ticket_models import (
+    CleanTicketDTO,
+    RawTicketDTO,
+    TicketType,
+    convert_ticket,
+)
+from .ticket_status import Impact, Priority, TicketStatus, Urgency
+
+__all__ = [
+    "TicketStatus",
+    "Priority",
+    "Urgency",
+    "Impact",
+    "TicketType",
+    "RawTicketDTO",
+    "CleanTicketDTO",
+    "convert_ticket",
+]
