+++ src/backend/models/ticket_models.py
